+++ node_io/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 1.0.0-dev.7.0
+
+- Fixed: handling errors in `Directory.delete`.
+- Fixed: handling relative paths in `Directory.list`.
+- Fixed: strong mode issues in `Directory` and `File`.
+- Added: `File.create`, `File.delete`, `File.readAsString`, `File.rename`,
+    `File.writeAsBytes`, `File.writeAsString`.
+- Fixed: handling errors in `File.open`, `File.stat`, `File.statSync`.
+- Fixed: converting byte data in `NodeIOSink`.
+- Added: Minimal scaffold for Link FS entities.
+
 ## 1.0.0-dev.6.0
 
 - Upgraded to latest build_node_compilers.